refactor: move parse_range and RefRange to src/git/refs.rs
Makes parse_range and RefRange pub(crate) so both the CLI subcommands
and the upcoming shim handlers can use them without duplicating logic.
Tests moved to src/git/refs.rs alongside the production code.
