+++ src/git/mod.rs
@@ -2,4 +2,5 @@ pub mod depfiles;
 pub mod diff;
 pub mod generated;
 pub mod reader;
+pub(crate) mod refs;
 pub mod worktree;
